d3d call audit + bgfx port. ff emulator shaders, state translation, draw submit, imgui + libmpv playback all through bgfx. crt post process. metal on macos, vulkan on linux + windows. ci runs all four platforms. compiled shaders ship in-tree under Data/shaders/.

Cleaned up D3D calls from the shim we will never use.
Fixed some weird bugs (green screen leaving media) that happened on GL, shouldnt happen anymore on Windows builds.
Linux and Windows now use Vulkan through BGFX. Significant performance increas.
This tree was being worked on while main was being patched I tried to ensure everything was covered, some outlier things may have leaked into the code (e.g I had to fix a pathing bug I had already fixed)

+++ theseus/desktop/launch.h
@@ -24,7 +24,7 @@
 //                     registered handler for the steam scheme, which
 //                     causes "Unknown protocol 'steam'" errors.
 //                     macOS / Windows: fall through to the system URL
-//                     opener -- it works there.
+//                     opener. it works there.
 //   - "<scheme>://..." system URL opener (xdg-open / open / ShellExecute)
 //   - anything else   pass to the shell as a command line, so the user
 //                     can paste a raw executable path, a path with args,
@@ -55,4 +55,4 @@ bool         LaunchOverlay_IsActive();
 float        LaunchOverlay_Alpha();
 void         LaunchOverlay_Tick();
 void         LaunchOverlay_Reset();
-unsigned int LaunchOverlay_LogoGLTex(int* outW, int* outH); // 0 if not loaded yet
+unsigned long long LaunchOverlay_LogoGLTex(int* outW, int* outH); // 0 if not loaded yet
